Refactored ketchup-combo into kup-combo
