+++ packages/ketchup/src/components/kup-combo/kup-combo.scss
@@ -32,7 +32,7 @@ $lateral-padding: 6px;
   z-index: 1; // [1]
 }
 
-.ketchup-combo {
+.kup-combo {
   // The div containing the input element.
   &__container {
     background-color: #ffffff;
