+++ packages/ketchup/src/components/kup-combo/readme.md
@@ -1,4 +1,4 @@
-# ketchup-combo
+# kup-combo
