+++ packages/ketchup/src/index.html
@@ -75,7 +75,7 @@
             is-clearable
             initial-value="ciao"
         ></ketchup-text-input>
-        <ketchup-combo is-clearable use-portal="true"></ketchup-combo>
+        <kup-combo is-clearable use-portal="true"></kup-combo>
         <h4>FLD 1</h4>
         <ketchup-fld id="kfld-1"></ketchup-fld>
         <h4>FLD 2</h4>
@@ -131,7 +131,7 @@
                     once = false;
 
                     // The combo box
-                    const comboBox = document.querySelector('ketchup-combo');
+                    const comboBox = document.querySelector('kup-combo');
                     comboBox.initialValue = {
                         id: 'Darth Vader',
                     };
